refactor(eval): move rank metrics to src/core/eval/rank-metrics.ts (R5 prep)
The collapse/churn detector shares nDCG/recall/MRR with the curate bench
and CI golden test; src/ must not import from scripts/, so the module
moves and scripts/akm-eval/src/curate-metrics.ts becomes a re-export
shim. No behavior change; curate-golden + curate-metrics tests green
through the shim.

Ref: docs/design/improve-collapse-churn-detector-design.md §2.4
